AI-Connector/requirements.txt
langchain-openai
openai
python-dotenv
flask
